Make test models with app_label 'ignore', not to trigger migrations, useful when we need to compose a test model that we don't want to migrate, for example when testing the sql produced by some of its field config

+++ tests/conftest.py
@@ -40,9 +40,10 @@ def clean_database(request):
     ]
 
     for model in models:
-        with connection.cursor() as cursor:
-            cursor.execute(f"DELETE FROM {model._meta.db_table}")
-            cursor.execute(f"REFRESH TABLE {model._meta.db_table}")
+        if model._meta.app_label != 'ignore':
+            with connection.cursor() as cursor:
+                cursor.execute(f"DELETE FROM {model._meta.db_table}")
+                cursor.execute(f"REFRESH TABLE {model._meta.db_table}")
 
 
 @pytest.fixture(scope="function", autouse=False)
